fix(sec): int-overflow casts, workflow read-only perms, dashboard XSS escaping (#1561)
- cpp/integer-multiplication-cast-to-long: (size_t) casts on the first
  factor of tensor-dim products feeding vector sizes / memcpy sizes /
  pointer indexes; (double) casts on float products feeding double
  accumulators — 37 in-repo files, 212 instances (969 more live in
  third_party/ git submodules — need upstream fixes, out of scope).
- actions/missing-workflow-permissions: analytics.yml gets top-level
  permissions: contents: read.
- js/xss-through-dom: esc() helper in site/dashboard/app.js; all
  message / e.message / k.key / savedKey interpolations in innerHTML
  sinks escaped (11 sinks incl. 5 same-pattern siblings).
- Dismissed (already, with reason): js/clear-text-storage ×2 — the
  single-user localhost dashboard intentionally keeps its Jarvis API
  key in localStorage (no server session exists).

Verification: g++ -fsyntax-only clean on all checkable files (remaining
need HIP/npu-SDK headers, same at HEAD); node --check passes.

+++ npu-infer/src/model.c
@@ -82,7 +82,7 @@ int npu_dequant_block(void* out, const void* in,
         num_rows = (int)config->npu_block_rows;
     
     // Zero entire output block
-    memset(bf16_out, 0, num_rows * block_cols * 2);
+    memset(bf16_out, 0, (size_t)num_rows * block_cols * 2);
     
     // For each BF16 value, read 2 consecutive I8 bytes as little-endian BF16
     // I8 data layout: [lo_byte_0, hi_byte_0, lo_byte_1, hi_byte_1, ...] per row
